+++ app/(kahlua)/login/info/page.tsx
@@ -3,9 +3,9 @@
 import * as React from 'react';
 import { useEffect } from 'react';
 import { useState } from 'react';
-import LoginModal from '@/components/kahluaLogin/loginModal';
-import LoginSelectBox from '@/components/kahluaLogin/LoginSelectBox';
-import NameInput from '@/components/kahluaLogin/nameInput';
+import LoginModal from '@/components/login/loginModal';
+import LoginSelectBox from '@/components/login/LoginSelectBox';
+import NameInput from '@/components/login/nameInput';
 
 // 기수 23 ~ 1기
 const generations: string[] = Array.from(
